small code refactoring
Document data_func decorator, and move to data_handler.py, where it is used.  Delete now-empty decorators.py.
