+++ model/tla.py
@@ -11,8 +11,7 @@ import pandas as pd
 from model import dd
 from model.metaclass_cache import MetaclassCache
 
-from model.data_handler import DataHandler
-from model.decorators import data_func
+from model.data_handler import DataHandler, data_func
 
 def tla_per_region(land_dist, custom_world_values=None):
     """
